Added graphql mutation

+++ src/main/kotlin/com/erwindev/openpayment/Application.kt
@@ -39,7 +39,7 @@ class Application: CommandLineRunner{
         csvReader.forEach{ line ->
             var openPayment: OpenPayment = OpenPayment(providerId = line[5],
                     providerName = line[6] + " " + line[7] + " " + line[8],
-                    paymentAmount = line[30].toDouble(),
+                    paymentAmount = line[30].toFloat(),
                     payerId = line[26],
                     payerName = line[27])
